Rename player.rs to player_plugin.rs
